+++ src/__tests__/index.test.ts
@@ -1,4 +1,4 @@
-import { run } from '../src/index'
+import { run } from '../index'
 import * as core from '@actions/core'
 import dedent from 'dedent-js'
