+++ tsconfig.json
@@ -10,5 +10,5 @@
     "esModuleInterop": true,
     "skipLibCheck": true
   },
-  "exclude": ["dist", "node_modules", "__tests__"]
+  "exclude": ["dist", "node_modules"]
 }
